+++ crates/client/sync/src/fetch/mod.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 use dc_db::DeoxysBackend;
 use dp_utils::{channel_wait_or_graceful_shutdown, wait_or_graceful_shutdown};
 use futures::prelude::*;
-use starknet_core::types::StarknetError;
+use starknet_core::types::{Felt, StarknetError};
 use starknet_providers::{ProviderError, SequencerGatewayProvider};
 use tokio::sync::{mpsc, oneshot};
 
@@ -14,6 +14,7 @@ use crate::l2::L2SyncError;
 
 pub mod fetchers;
 
+#[allow(clippy::too_many_arguments)]
 pub async fn l2_fetch_task(
     backend: Arc<DeoxysBackend>,
     first_block: u64,
@@ -22,6 +23,7 @@ pub async fn l2_fetch_task(
     provider: Arc<SequencerGatewayProvider>,
     sync_polling_interval: Option<Duration>,
     once_caught_up_callback: oneshot::Sender<()>,
+    chain_id: Felt,
 ) -> anyhow::Result<()> {
     // First, catch up with the chain
     let backend = &backend;
@@ -32,7 +34,7 @@ pub async fn l2_fetch_task(
         // Fetch blocks and updates in parallel one time before looping
         let fetch_stream = (first_block..).take(n_blocks_to_sync.unwrap_or(u64::MAX) as _).map(|block_n| {
             let provider = Arc::clone(&provider);
-            async move { (block_n, fetch_block_and_updates(backend, block_n, provider).await) }
+            async move { (block_n, fetch_block_and_updates(backend, block_n, provider, chain_id).await) }
         });
 
         // Have 10 fetches in parallel at once, using futures Buffered
@@ -67,7 +69,7 @@ pub async fn l2_fetch_task(
         interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
         while wait_or_graceful_shutdown(interval.tick()).await.is_some() {
             loop {
-                match fetch_block_and_updates(backend, next_block, Arc::clone(&provider)).await {
+                match fetch_block_and_updates(backend, next_block, Arc::clone(&provider), chain_id).await {
                     Err(L2SyncError::Provider(ProviderError::StarknetError(StarknetError::BlockNotFound))) => {
                         break;
                     }