+++ crates/client/sync/src/l2.rs
@@ -151,8 +151,9 @@ async fn l2_verify_and_apply_task(
         .await?;
 
         let sw = PerfStopwatch::new();
-        backend.maybe_flush()?;
-        stopwatch_end!(sw, "flush db: {:?}");
+        if backend.maybe_flush()? {
+            stopwatch_end!(sw, "flush db: {:?}");
+        }
 
         log::info!(
             "✨ Imported #{} ({}) and updated state root ({})",
@@ -341,6 +342,7 @@ pub async fn sync(
         Arc::clone(&provider),
         config.sync_polling_interval,
         once_caught_up_cb_sender,
+        chain_id,
     ));
     join_set.spawn(l2_block_conversion_task(fetch_stream_receiver, block_conv_sender, chain_id));
     join_set.spawn(l2_verify_and_apply_task(